Loggers weren't logging if they overrode the base class log(int state) - because it had been changed to log(long state). Put a final int version in the base class to prevent overriding.

+++ trunk/src/dr/evomodel/operators/ImportanceNarrowExchange.java
@@ -254,9 +254,10 @@ public class ImportanceNarrowExchange extends AbstractTreeOperator implements Tr
         justAccepted = true;
     }
 
-    private final int lFreq = 1000;
-    private int lastLog = -lFreq-1;
-    public boolean logNow(int state) {
+    private final long lFreq = 1000;
+    private long lastLog = -lFreq-1;
+
+    public boolean logNow(long state) {
 
         boolean r = justAccepted;
         if( lastLog + lFreq >= state ) {
